timescaler works but is actually one giant hack

+++ SeadsFront/seadssite/views.py
@@ -1,6 +1,9 @@
-from .models import Devices, Map
+import time
+import requests
+import re
+import ast
+import json
 from django.http import HttpResponseRedirect, HttpResponse
-from seadssite.forms import UserForm, UserProfileForm
 from django.template import RequestContext
 from django.shortcuts import render, render_to_response
 from django.contrib.auth import authenticate, login
@@ -8,9 +11,9 @@ from django.views.generic import View, CreateView
 from django.core.urlresolvers import reverse
 from django.views.generic.base import TemplateView
 from django.contrib.auth.models import User
-import requests
-import re
-import ast
+from seadssite.forms import UserForm, UserProfileForm
+from .models import Devices, Map
+
 
 from django.core.mail import send_mail
 from django.shortcuts import render_to_response as render_to
@@ -98,16 +101,14 @@ def DashboardView(request):
     current_user = request.user    
     user_devices_map = Map.objects.filter(user=current_user.id)
 
-    #if the user clicked device register
+    #if the user clicked register
     if request.POST.get('register'):
         #get the new device ID and name from the post
         new_device_id = request.POST.get('device_id')
         new_device_name = request.POST.get('device_name')
-        #check if device is already registered
-        if Devices.objects.filter(device_id=new_device_id):
-            '''
-            TODO: Deal with this for the user
-            '''
+        #check if device is already registered to this user
+        D = Devices.objects.filter(device_id=new_device_id)
+        if Map.objects.filter(user=current_user.id, device=D):
             alerts.append("The device you've attempted to register has already been registered.")
             print "Matching device ID"
         else:
@@ -122,16 +123,25 @@ def DashboardView(request):
             except TypeError:
                 print "Invalid Device ID"
 
-    #if the user clicked device delete
+    #if the user clicked delete
     elif request.POST.get('delete'):
         #get the device ID from the post
-        '''
-        TODO: Verify that device ID posted for deletion is associated with current user
-        this avoids sending a post request to delete random peoples devices
-        '''
         device_id = request.POST.get('delete')
-        #delete the record in the DB (cascades to delete the map)
-        Devices.objects.filter(device_id = device_id).delete()
+        #get respective DB objects
+        D = Devices.objects.filter(device_id = device_id)
+        M = Map.objects.filter(device=D)
+        #if the user owns the device they are trying to delete
+        if Map.objects.filter(user=current_user.id, device=D):
+            #if multiple users own the device
+            if len(M) > 1:
+                #just delete this users map to the device
+                Map.objects.filter(user=current_user.id, device=D).delete()
+            else:
+                #delete the device itself (cascades to map)
+                Devices.objects.filter(device_id = device_id).delete()
+        else:
+            print "you don't own the device you're deleting, or it doesn't exist"
+
 
     return render(request, 'dashboard.html', {'maps': user_devices_map, 'alerts':alerts})
 
@@ -217,72 +227,28 @@ hit api asking for all data for a device (this gets displayed as soon as the
   user hits the page) then give options for different api calls
 '''
 def VisualizationView(request, device_id):
-  api_string = "http://128.114.59.76:8080/{}".format(device_id)
-  api_string += "?type={}".format("I")
-  api_string += "&start_time=1416527462&end_time=141652754"
-  r = requests.get("http://128.114.59.76:8080/1?type=I&start_time=1416527462&end_time=1416527542")
-  #r = requests.get(api_string)
-  r1 = re.sub("\]\n\[", "],\n[", r.text)
-  r2 = re.sub("\]\n\]", "],\n]", r1)
-  omlett = ast.literal_eval(r2)
-  print omlett
-  for row in omlett:
-    if row[1] != "I":
-        row[1] = int(row[1])
-  print omlett
-  fake_data = omlett
-  if(request.POST.get('all')):
-    api_string = "DB/{}".format(device_id)
-    fake_data = [
-    ['Time', 'KW/H', 'Temp'],
-    ['1',  50,      70],
-    ['2',  60,      77],
-    ['3',  80,      82],
-    ['4',  50,      76],
-    ['5',  50,      70],
-    ['6',  50,      70],
-    ['7',  60,      77],
-    ['8',  80,      82],
-    ['9',  50,      76],
-    ['10',  50,      70],
-    ['11',  50,      70],
-    ['12',  60,      77],
-    ['13',  80,      82],
-    ['14',  50,      76],
-    ['15',  50,      70],
-    ['16',  50,      70],
-    ['17',  60,      77],
-    ['18',  80,      82],
-    ['19',  50,      76],
-    ['20',  50,      70]
-    ]
+    api_string = "http://128.114.59.76:8080/{}".format(device_id)
+    api_string += "?type={}".format("W")    
+    #print request
+    params = request.GET
+    start_time = params.get('start_time', 0)
+    end_time = params.get('end_time', int(time.time()))
+    api_string += "&start_time={}&end_time={}".format(start_time, end_time)
+    api_response = requests.get(api_string).text
+    api_response = ast.literal_eval(api_response)
+    for row in api_response:
+        for index, value in enumerate(row):
+            if index > 0 and value not in ['I','W']:
+                row[index] = int(value)
 
-  elif(request.POST.get('month')):
-    api_string = api_string + "/month"
-    fake_data = [
-    ['Time', 'KW/H', 'Temp'],
-    ['1',  50,      70],
-    ['2',  60,      77],
-    ['3',  80,      82],
-    ['4',  50,      76],
-    ['5',  50,      70],
-    ['6',  50,      70],
-    ['7',  60,      77],
-    ['8',  80,      82],
-    ['9',  50,      76],
-    ['10',  50,      70]
-    ]
+    s =json.dumps(api_response)
+    goodness = json.loads(s)
+    print goodness
 
-  elif(request.POST.get('week')):
-    api_string = api_string + "/week"
-    fake_data = [
-    ['Time', 'KW/H', 'Temp'],
-    ['1',  50,      70],
-    ['2',  60,      77],
-    ['3',  80,      82],
-    ['4',  50,      76],
-    ['5',  50,      70]
-    ]
+    if request.is_ajax():
+        a = []
+        a.append(api_response)
+        return HttpResponse(json.dumps(goodness), content_type="application/json")
 
-  return render(request, 'visualization.html', {'data':fake_data, 'api_call':api_string})
+    return render(request, 'visualization.html', {'data':api_response, 'api_call':api_string})
